pyproject.toml
[project]
name = "reserva"
version = "0.1.0"
description = ""
authors = [{ name = "tomas.rajao@weduu.com" }]
readme = "README.md"
requires-python = ">=3.11, <4.0"
dependencies = ["fastapi[standard] (>=0.115.8,<0.116.0)", "sqlalchemy (>=2.0.38,<3.0.0)", "pydantic-settings (>=2.7.1,<3.0.0)", "alembic (>=1.14.1,<2.0.0)", "pyjwt (>=2.10.1,<3.0.0)", "pwdlib[argon2] (>=0.2.1,<0.3.0)"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
taskipy = "^1.14.1"
ruff = "^0.9.4"

[tool.ruff]
line-length = 120
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_opitions]
pythonpath = "."
addopts = '-p no:warnings'



[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
formatf = 'ruff format'
run = 'fastapi dev reserva/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=reserva -vv'
post_test = 'coverage html'
makemigrations = 'alembic revision --autogenerate -m '
migrate = 'alembic upgrade head'